plugins/StashWebpRemover/requirements.txt
Pillow
stashapi